+++ packages/memory-store/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @melandlabs/memory-store
 
+## 0.1.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @melandlabs/rag@0.1.5
+
 ## 0.1.5
 
 ### Patch Changes